+++ checks/src/main/kotlin/com/faithlife/lint/IssueRegistry.kt
@@ -14,6 +14,7 @@ class IssueRegistry : ApiIssueRegistry() {
 
     override val issues = listOf(
         FiniteWhenCasesDetector.ISSUE,
+        ForEachFunctionDetector.ISSUE,
         IndirectSuperCallDetector.ISSUE,
         ObsoleteDateInstantiationDetector.ISSUE,
         RedundantCoroutineScopeDetector.ISSUE,